refactor app.js - added topMenu and SideMenu
