+++ client/src/components/menus/topMenu.js
@@ -2,9 +2,9 @@ import React, { Component } from 'react';
 import { Route, NavLink } from 'react-router-dom';
 import styled from 'styled-components';
 
-const NavDiv = styled.div`
-  justify-content: space-evenly;
-`;
+// const NavDiv = styled.div`
+//   justify-content: space-evenly;
+// `;
 
 const TopMenu = () => {
 
@@ -12,13 +12,13 @@ const TopMenu = () => {
 
 
     return (
-        <NavDiv>
+        <div className="topMenu">
 
         <NavLink to="/signup">Sign Up</NavLink>
         <NavLink to="/signin">Sign In</NavLink>
 
         <NavLink to="/signout">Sign Out</NavLink>
-      </NavDiv>
+      </div>
     );
   }
